feat(forms/enumeration): items with checkboxes
