+++ packages/forms/src/widgets/EnumerationWidget/EnumerationWidget.actions.js
@@ -1,0 +1,53 @@
+export const validate = {
+	disabled: true,
+	label: 'Validate',
+	icon: 'talend-check',
+	id: 'validate',
+	key: 'validate',
+};
+
+export const abort = {
+	label: 'Abort',
+	icon: 'talend-cross',
+	id: 'abort',
+	key: 'abort',
+};
+
+export const loading = {
+	label: 'Loading',
+	icon: 'talend-cross',
+	inProgress: true,
+	id: 'loading',
+};
+
+export const edit = {
+	disabled: false,
+	label: 'Edit',
+	icon: 'talend-pencil',
+	id: 'edit',
+};
+
+export const removeOne = {
+	label: 'Remove value',
+	icon: 'talend-trash',
+	id: 'delete',
+};
+
+export const add = {
+	label: 'Add item',
+	icon: 'talend-plus',
+	id: 'add',
+};
+
+export const search = {
+	disabled: false,
+	label: 'Search for specific values',
+	icon: 'talend-search',
+	id: 'search',
+};
+
+export const remove = {
+	label: 'Remove selected values',
+	icon: 'talend-trash',
+	id: 'delete',
+};
