+++ packages/forms/src/widgets/EnumerationWidget/EnumerationWidget.constants.js
@@ -1,0 +1,17 @@
+export const DISPLAY_MODE_EDIT = 'DISPLAY_MODE_EDIT';
+export const DISPLAY_MODE_SELECTED = 'DISPLAY_MODE_SELECTED';
+export const ENUMERATION_SEARCH_ACTION = 'ENUMERATION_SEARCH_ACTION';
+export const ENUMERATION_ADD_ACTION = 'ENUMERATION_ADD_ACTION';
+export const ENUMERATION_REMOVE_ACTION = 'ENUMERATION_REMOVE_ACTION';
+export const ENUMERATION_RENAME_ACTION = 'ENUMERATION_RENAME_ACTION';
+export const ENUMERATION_RESET_LIST = 'ENUMERATION_RESET_LIST';
+
+export const DISPLAY_MODE_DEFAULT = 'DISPLAY_MODE_DEFAULT';
+export const DISPLAY_MODE_ADD = 'DISPLAY_MODE_ADD';
+export const DISPLAY_MODE_SEARCH = 'DISPLAY_MODE_SEARCH';
+
+export const DISPLAY_MODE_CHECKBOX = 'DISPLAY_MODE_CHECKBOX';
+
+export const ITEMS_DEFAULT_HEIGHT = 33;
+
+export const DUPLICATION_ERROR = 'This term is already in the list';
