+++ packages/forms/stories/json/enumerationWithCheckboxes.json
@@ -1,0 +1,45 @@
+{
+  "jsonSchema": {
+    "title": "A enumeration form",
+    "description": "A enumeration form example.",
+    "type": "object",
+    "properties": {
+      "documents": {
+        "type": "array",
+        "required": true,
+        "allowDuplicates": true,
+        "displayMode": "DISPLAY_MODE_CHECKBOX",
+        "items": {
+          "type": "object",
+          "properties": {
+            "values": {
+              "type": "array",
+              "items": {
+                "type": "string"
+              }
+            }
+          }
+        }
+      }
+    }
+  },
+  "uiSchema": {
+    "documents": {
+      "ui:widget": "enumeration"
+    }
+  },
+  "properties": {
+    "documents": [
+      {
+        "values": [
+          "France"
+        ]
+      },
+      {
+        "values": [
+          "Espagne"
+        ]
+      }
+    ]
+  }
+}
